refactor: separate BodyInfo service logic from UserService into BodyInfoService.
