+++ src/main/java/linkfit/controller/BodyInfoController.java
@@ -1,0 +1,2 @@
+package linkfit.controller;public class BodyInfoController {
+}
